+++ sdk/core/examples/get_hyper_block_by_nonce.rs
@@ -1,4 +1,4 @@
-use multiversx_sdk::blockchain::{GatewayProxy, DEVNET_GATEWAY};
+use multiversx_sdk::gateway::{GatewayProxy, DEVNET_GATEWAY};
 
 #[tokio::main]
 async fn main() {
