+++ sdk/core/examples/sign_tx.rs
@@ -1,6 +1,6 @@
 use multiversx_sdk::{
-    blockchain::{GatewayProxy, DEVNET_GATEWAY},
     data::transaction::Transaction,
+    gateway::{GatewayProxy, DEVNET_GATEWAY},
     wallet::Wallet,
 };
