+++ sdk/core/src/gateway/gateway_block.rs
