+++ sdk/core/src/gateway/gateway_network.rs
