+++ sdk/core/src/gateway/gateway_proxy.rs
