+++ sdk/core/src/gateway/gateway_tx_retrieve.rs
