+++ sdk/core/src/lib.rs
@@ -1,5 +1,5 @@
-pub mod blockchain;
 pub mod crypto;
 pub mod data;
+pub mod gateway;
 pub mod utils;
 pub mod wallet;
